Fix more simple warnings

+++ src/Agent.cpp
@@ -53,7 +53,7 @@ void Agent::printSelf() {
   printf("Agent age=%i\n", age);
   printf("\ttemp pref=%f\n", temperature_preference);
 
-  for (int i = 0; i < mutations.size(); i++) {
+  for (size_t i = 0; i < mutations.size(); i++) {
     cout << mutations[i];
   }
   cout << endl;
